Fix issue #1
Changed the way the file filling works by filling the entire file with a random byte, not quite pattern filling BUT it works extremely well and fast.

also fixed an issue where zero would hide as an iteration. may switch from iterative for loop in `main.rs` to an .iter() function.